% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beispieldatensatzdoc.R
\docType{data}
\name{beispieldatensatz}
\alias{beispieldatensatz}
\title{Beispiel Datensatz}
\format{
Es gibt 4 Attributen: \code{product_ndc} ist die Identifikationsnummern von Medikamenten;
\code{finished} beschreibt ob das Medikament fertig entwickelt oder nicht;
\code{dosage_form} beschreibt die Kategorie von dem Medikament;
\code{marketing_start_date} beschreibt die Anfangsdatum von der Vermarktung.
}
\usage{
beispieldatensatz
}
\description{
Ein Datensatz bestehend aus 10 Tupels mit 4 Spalten.
}
\keyword{datasets}
