% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{Berechne die numerische Zusammenfassung}
\usage{
calculate(data = getFromAPI())
}
\arguments{
\item{data}{die Daten, auf den die Berechnung basiert.}
}
\value{
ein tibble von der numerische Zusammenfassung.
}
\description{
\code{calculate} berechnet die numerische Zusammenfassung von \code{data}.
}
\details{
Dies ist eine Funktion, die die median von marketing_start_date berechnet in
den Untergruppen von finished und dosage_form.
}
\examples{
calculate()
calculate(getFromAPI())
}
