% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawGraph.R
\name{drawGraph}
\alias{drawGraph}
\title{Zeichne die Grafik}
\usage{
drawGraph(data = getFromAPI())
}
\arguments{
\item{data}{die Daten, auf den die Grafik basiert.}
}
\value{
void.
}
\description{
\code{drawGraph} zeichnet die Grafik, die die eingegebene Daten benutzt.
}
\details{
Dies ist eine Funktion, die eine Grafik zeichnet, die Informationen ueber
die Relation zwischen count und marketing starting date darstellt.
Die Quelle ist \code{data}. Diese Funktion haengt von \code{\link[ggplot2]{ggplot}}
und \code{\link[parsedate]{parse_date}} ab.
}
\examples{
drawGraph()
drawGraph(getFromAPI())
}
