% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromAPI.R
\name{getFromAPI}
\alias{getFromAPI}
\title{Herunterlade Daten von API}
\usage{
getFromAPI(startYear = 2015, endYear = 2018, amount = 500)
}
\arguments{
\item{startYear}{Das erste Jahr}

\item{endYear}{Das letzte Jahr}

\item{amount}{Die Anzahl von Faellen, maximum 1000}
}
\value{
Ein data frame als tibble, das die Daten von \code{startYear} bis
\code{endYear} enthaelt. Die Anzahl von Daten ist \code{amount}.
}
\description{
\code{getFromAPI} Herunterlaedt die Daten von FDA von den Medikamenten in
den USA.
}
\details{
Dies ist eine Funktion, die riesige Daten von allen Medikamenten in den USA
 liefert. Die Quelle ist \url{https://api.fda.gov/drug/ndc.json}. Diese
 Funktion haengt von \code{\link[httr]{GET}}
und \code{\link[jsonlite]{fromJSON}} ab.
}
\examples{
getFromAPI()
getFromAPI(2005, 2015, 500)
}
